Move health into dns/health
The contents of the dns package will all be moved to the new dns controller component as top level packages:

```
.
├── controller
└── pkg
    ├── health
    └── provider
        ├── aws
        ├── fake
        └── google
```
