+++ pkg/controllers/dnshealthcheckprobe/notifier.go
@@ -10,7 +10,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/Kuadrant/multicluster-gateway-controller/pkg/apis/v1alpha1"
-	"github.com/Kuadrant/multicluster-gateway-controller/pkg/health"
+	"github.com/Kuadrant/multicluster-gateway-controller/pkg/dns/health"
 )
 
 type StatusUpdateProbeNotifier struct {
